fastboot: fix offset truncation in oem_read for >4GB access
The offset parameter in oem_read() was defined as u32, causing
truncation when reading data beyond 4GB. This affected the entire
read path through fastboot_mtd_flash_read(), fastboot_blk_read(),
and fastboot_mmc_read().

Changes:
- Change offset type from u32 to u64 in oem_read()
- Update offset parameter to u64 in all read function signatures:
  * _fb_mtd_read()
  * fastboot_mtd_flash_read()
  * fastboot_blk_read()
  * fastboot_mmc_read()
- Use simple_strtoull() instead of simple_strtoul() for parsing
- Fix printf format specifiers from %x/%lx to %llx for 64-bit offsets

This enables correct reading of data from storage devices with
capacities exceeding 4GB.

+++ include/fb_blk.h
@@ -46,7 +46,7 @@ void fastboot_blk_erase(const char *cmd, char *response);
  * @part: Named partition to erase
  * @response: Pointer to fastboot response buffer
  */
-u32 fastboot_blk_read(const char *part, u32 offset,
+u32 fastboot_blk_read(const char *part, u64 offset,
 			void *download_buffer, char *response);
 
 #endif
